@@ -6,7 +6,7 @@ from setuptools import find_packages, setup
 setup(
     name="pwned-passwords-django",
     zip_safe=False,  # eggs are the devil.
-    version="1.5",
+    version="1.6a1",
     description="A Pwned Passwords implementation for Django sites.",
     long_description=open(os.path.join(os.path.dirname(__file__), "README.rst")).read(),
     author="James Bennett",
@@ -18,20 +18,19 @@ setup(
         "Development Status :: 5 - Production/Stable",
         "Environment :: Web Environment",
         "Framework :: Django",
-        "Framework :: Django :: 2.2",
-        "Framework :: Django :: 3.1",
         "Framework :: Django :: 3.2",
+        "Framework :: Django :: 4.0",
         "Intended Audience :: Developers",
         "License :: OSI Approved :: BSD License",
         "Operating System :: OS Independent",
         "Programming Language :: Python",
         "Programming Language :: Python :: 3",
-        "Programming Language :: Python :: 3.6",
         "Programming Language :: Python :: 3.7",
         "Programming Language :: Python :: 3.8",
         "Programming Language :: Python :: 3.9",
+        "Programming Language :: Python :: 3.10",
         "Topic :: Utilities",
     ],
-    python_requires=">=3.6",
-    install_requires=["Django>=2.2,!=3.0.*", "requests"],
+    python_requires=">=3.7",
+    install_requires=["Django>=3.2", "requests"],
 )
