@@ -7,9 +7,9 @@ extensions = ["sphinx.ext.intersphinx"]
 source_suffix = ".rst"
 master_doc = "index"
 project = "pwned-passwords-django"
-copyright = "2018-2021, James Bennett"
-version = "1.5"
-release = "1.5"
+copyright = "2018-2022, James Bennett"
+version = "1.6"
+release = "1.6a1"
 exclude_trees = ["_build"]
 pygments_style = "sphinx"
 htmlhelp_basename = "pwned-passwords-djangodoc"
